Allow for the CM version to be specified as the second argument.

+++ install_clouderamanageragent.sh
@@ -19,12 +19,18 @@ if [ -z "$SCMHOST" ]; then
   echo "ERROR: Missing SCM hostname."
   exit 1
 fi
+SCMVERSION=$2
 if rpm -q redhat-lsb-core; then
   OSREL=`lsb_release -rs | awk -F. '{print $1}'`
 else
   OSREL=`rpm -qf /etc/redhat-release --qf="%{VERSION}\n"`
 fi
-wget -q http://archive.cloudera.com/cm5/redhat/${OSREL}/x86_64/cm/cloudera-manager.repo -O /etc/yum.repos.d/cloudera-manager.repo
+if [ ! -f /etc/yum.repos.d/cloudera-manager.repo ]; then
+  wget -q http://archive.cloudera.com/cm5/redhat/${OSREL}/x86_64/cm/cloudera-manager.repo -O /etc/yum.repos.d/cloudera-manager.repo
+  if [ -n "$SCMVERSION" ]; then
+    sed -e "s|/cm/5/|/cm/${SCMVERSION}/|" -i /etc/yum.repos.d/cloudera-manager.repo
+  fi
+fi
 yum -y -e1 -d1 install cloudera-manager-agent
 sed -i -e "/server_host/s|=.*|=${SCMHOST}|" /etc/cloudera-scm-agent/config.ini
 service cloudera-scm-agent start
